requirements.txt
Flask>=2.2.5,<2.3
mysql-connector-python
python-dotenv